+++ vetclinic-server/src/main/kotlin/personal/ciai/vetclinic/service/AdministrativeService.kt
@@ -11,14 +11,14 @@ import personal.ciai.vetclinic.repository.AdministrativeRepository
 @Service
 class AdministrativeService(@Autowired val adminRepository: AdministrativeRepository) {
 
-    fun existsById(id: Long): Boolean = adminRepository.existsById(id)
+    fun existsById(id: Int): Boolean = adminRepository.existsById(id)
 
     fun getAllAdministrative(): List<AdministrativeDTO> = adminRepository.findAll()
         .map { it.toDTO() }
 
-    fun getAdministrativeById(id: Long): AdministrativeDTO = getAdministrativeEntity(id).toDTO()
+    fun getAdministrativeById(id: Int): AdministrativeDTO = getAdministrativeEntity(id).toDTO()
 
-    private fun getAdministrativeEntity(id: Long): Administrative = adminRepository.findById(id)
+    private fun getAdministrativeEntity(id: Int): Administrative = adminRepository.findById(id)
         .orElseThrow { NotFoundException("Administrative account with Id $id not found") }
 
     fun save(adminDTO: AdministrativeDTO) {
@@ -27,13 +27,13 @@ class AdministrativeService(@Autowired val adminRepository: AdministrativeReposi
         else throw ConflictException("Administrative account with Id ${adminDTO.id} already exist")
     }
 
-    fun update(adminDTO: AdministrativeDTO, id: Long) {
+    fun update(adminDTO: AdministrativeDTO, id: Int) {
         val admin: Administrative = getAdministrativeEntity(id)
 
         adminRepository.save(Administrative.fromDto(adminDTO, admin))
     }
 
-    fun delete(id: Long) {
+    fun delete(id: Int) {
         val admin: Administrative = getAdministrativeEntity(id)
 
         adminRepository.delete(admin)
